zillow_wf/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
sqlmodel==0.0.14
alembic==1.12.1
psycopg2-binary==2.9.9
httpx==0.25.2
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==0.19.2
typer==0.9.0
beautifulsoup4==4.12.2
lxml==4.9.3
jupyter==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
backoff==2.2.1
requests==2.31.0

# Additional dependencies for flexible_waterfront_extractor.py
tqdm==4.66.1
asyncio-mqtt==0.16.1

